requirements.txt
google-cloud-pubsub==1.3.0
google-cloud-dlp==0.13.0
google-cloud-storage==1.27.0
